Switched XHR to sync  !!
This is a horrible idea.  I have to build the assetloader properly or the animations may not work since they rely on the JSON being loaded and I haven't made the xhr status stop the game from running until it's loaded so I just made it stop the whole page.  It works, right?  Bad, haha.